services/candidature_service/requirements_depot_dossier.txt
# requirements_depot_dossier.txt
# Dépendances pour Sprint 2 - Dépôt de Dossier

# Django & DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# Authentication & Security
djangorestframework-authtoken>=1.4.0
django-guardian>=2.4.0
cryptography>=41.0.0

# Async & Task Queue
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
redis>=4.5.0

# File Processing & OCR
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-pptx>=0.6.21

# Database
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2

# Monitoring & Logging
django-structlog>=7.0.0
sentry-sdk>=1.39.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
coverage>=7.3.0
factory-boy>=3.3.0
faker>=20.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
pylint>=3.0.2
django-debug-toolbar>=4.2.0
drf-spectacular
